Refactor KML routines
