+++ monitoring/monitorlib/kml/__init__.py
@@ -1,0 +1,1 @@
+KML_NAMESPACE = {"kml": "http://www.opengis.net/kml/2.2"}
